Move type hash func from Utils to Script class
